+++ resources/views/filament/help-button.blade.php
@@ -1,0 +1,1 @@
+<x-filament::button tag="a" :href="$url" color="gray" size="sm" icon="heroicon-o-lifebuoy" outlined>Get help</x-filament::button>